% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R
\docType{package}
\name{coursemap-package}
\alias{coursemap-package}
\alias{coursemap}
\title{Course Dependency Map Visualization}
\description{
Course Map - R Interface
}
\details{
R wrapper for the Rust-based course-map tool.
Provides intuitive API with seamless integration with knitr/Quarto.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kenjisato/coursemap}
  \item Report bugs at \url{https://github.com/kenjisato/coursemap/issues}
}

}
\author{
Kenji Sato
}
\keyword{package}
