% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{generate_inline_svg}
\alias{generate_inline_svg}
\title{Generate SVG content as string for inline embedding}
\usage{
generate_inline_svg(input_dir, config_path)
}
\description{
Generate SVG content as string for inline embedding
}
