% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R, R/extendr-wrappers.R
\name{graphviz_available}
\alias{graphviz_available}
\title{Check if Graphviz is available}
\usage{
graphviz_available()

graphviz_available()
}
\value{
TRUE if Graphviz is available, FALSE otherwise
}
\description{
Check if Graphviz is available

Check if Graphviz is available
}
\examples{
\dontrun{
if (graphviz_available()) {
  message("Can generate PNG/SVG files")
} else {
  message("Only DOT format available")
}
}

}
