% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R, R/extendr-wrappers.R
\name{graphviz_info}
\alias{graphviz_info}
\title{Get Graphviz version information}
\usage{
graphviz_info()

graphviz_info()
}
\value{
Graphviz version string
}
\description{
Get Graphviz version information

Get Graphviz version information
}
\examples{
\dontrun{
if (graphviz_available()) {
  cat(graphviz_info())
}
}

}
