% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{parse_documents}
\alias{parse_documents}
\title{Parse documents in a directory and return metadata}
\usage{
parse_documents(input_dir, config_path)
}
\description{
Parse documents in a directory and return metadata
}
