% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R
\name{plot.coursemap}
\alias{plot.coursemap}
\title{Plot method for coursemap objects}
\usage{
\method{plot}{coursemap}(x, ...)
}
\arguments{
\item{x}{A coursemap object}

\item{...}{Additional arguments (ignored)}
}
\description{
Displays the course map in the current graphics device.
This method provides knitr/Quarto integration and environment-aware display.
}
\examples{
\dontrun{
cm <- coursemap("./courses")
plot(cm)  # Display in current environment
}

}
