% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coursemap.R
\docType{package}
\name{coursemap-package}
\alias{coursemap-package}
\title{Course Map - R Interface}
\description{
R wrapper for the Rust-based course-map tool.
Provides intuitive API with seamless integration with knitr/Quarto.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kenjisato/coursemap}
  \item Report bugs at \url{https://github.com/kenjisato/coursemap/issues}
}

}
\author{
\strong{Maintainer}: Kenji Sato \email{mail@kenjisato.jp}

}
\keyword{internal}
