% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{generate_course_map}
\alias{generate_course_map}
\title{Generate a course dependency map}
\usage{
generate_course_map(input_dir, output_path, format, config_path)
}
\description{
Generate a course dependency map
}
